+++ neuromation/cli/formatters/jobs.py
@@ -10,8 +10,9 @@ import humanize
 from click import style, unstyle
 
 from neuromation.api import JobDescription, JobStatus, JobTelemetry, Resources
+from neuromation.cli.parse_utils import JobColumnInfo
 from neuromation.cli.printer import StreamPrinter, TTYPrinter
-from neuromation.cli.utils import JobColumnInfo, format_size
+from neuromation.cli.utils import format_size
 
 from .ftable import table
